# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"WhatsappResponseModel": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"timestamp": {
								"type": "number",
								"format": "double"
							},
							"queueName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"timestamp",
							"queueName",
							"id"
						],
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"INotifyToWhatsappSchema": {
				"properties": {
					"numberPhone": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"numberPhone",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmailResponseModel": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"timestamp": {
								"type": "number",
								"format": "double"
							},
							"queueName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"timestamp",
							"queueName",
							"id"
						],
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"INotifyToEmailSchema": {
				"properties": {
					"senderList": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"subject": {
						"type": "string"
					},
					"htmlMessage": {
						"type": "string"
					}
				},
				"required": [
					"senderList",
					"subject",
					"htmlMessage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StatusCodes": {
				"enum": [
					100,
					101,
					102,
					103,
					200,
					201,
					202,
					203,
					204,
					205,
					206,
					207,
					300,
					301,
					302,
					303,
					304,
					305,
					307,
					308,
					400,
					401,
					402,
					403,
					404,
					405,
					406,
					407,
					408,
					409,
					410,
					411,
					412,
					413,
					414,
					415,
					416,
					417,
					418,
					419,
					420,
					421,
					422,
					423,
					424,
					426,
					428,
					429,
					431,
					451,
					500,
					501,
					502,
					503,
					504,
					505,
					507,
					511
				],
				"type": "number"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "tracking-notification",
		"version": "0.0.1",
		"description": "Tracking notification server with Express and TypeScript",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "maur025"
		}
	},
	"paths": {
		"/whatsapps/queue": {
			"post": {
				"operationId": "AddWhatsappToQueue",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WhatsappResponseModel"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INotifyToWhatsappSchema"
							}
						}
					}
				}
			}
		},
		"/emails/queue": {
			"post": {
				"operationId": "AddQueueEmail",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmailResponseModel"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INotifyToEmailSchema"
							}
						}
					}
				}
			}
		},
		"/channels": {
			"get": {
				"operationId": "GetChannels",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"data": {
														"properties": {
															"userParams": {
																"items": {
																	"properties": {
																		"default": {
																			"type": "string"
																		},
																		"type": {
																			"type": "string"
																		},
																		"description": {
																			"type": "string"
																		},
																		"field": {
																			"type": "string"
																		}
																	},
																	"required": [
																		"default",
																		"type",
																		"description",
																		"field"
																	],
																	"type": "object"
																},
																"type": "array"
															},
															"params": {
																"items": {
																	"properties": {
																		"value": {
																			"type": "string"
																		},
																		"type": {
																			"type": "string"
																		},
																		"field": {
																			"type": "string"
																		}
																	},
																	"required": [
																		"value",
																		"type",
																		"field"
																	],
																	"type": "object"
																},
																"type": "array"
															}
														},
														"required": [
															"userParams",
															"params"
														],
														"type": "object"
													},
													"deleted": {
														"type": "boolean"
													},
													"id": {
														"type": "string"
													},
													"cProtocol": {
														"properties": {
															"script": {
																"type": "string"
															},
															"deleted": {
																"type": "boolean"
															},
															"id": {
																"type": "string"
															},
															"name": {
																"type": "string",
																"enum": [
																	"mail_smtp",
																	"sms_service",
																	"wp_service"
																]
															}
														},
														"required": [
															"name"
														],
														"type": "object"
													},
													"cProtocolId": {
														"type": "string"
													},
													"name": {
														"type": "string"
													}
												},
												"required": [
													"cProtocol",
													"cProtocolId",
													"name"
												],
												"type": "object"
											},
											"type": "array"
										},
										"code": {
											"$ref": "#/components/schemas/StatusCodes"
										}
									},
									"required": [
										"data",
										"code"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api/notifications"
		}
	]
}